Migrate data type to floats

+++ astroimsumlib_c/include/astroimsumlib.h
@@ -23,7 +23,7 @@ int astroimsum_access_frame(
     struct astroimsum_frame* frame_in,
     int* width_out,
     int* height_out,
-    unsigned long** data_out);
+    float** data_out);
 
 int astroimsum_destroy_frame(
     struct astroimsum* handle_in, struct astroimsum_frame* frame_in);
